requirements.txt
# Lift OS Core - Shared Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=42.0.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# KSE Memory SDK
kse-memory-sdk==2.0.0

# Logging
python-json-logger==2.0.7
structlog==23.2.0

# Monitoring & Metrics
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# OAuth
authlib==1.2.1
python-jose[cryptography]==3.3.0

# Billing
stripe==7.8.0

# Secrets Management (Optional backends)
# boto3==1.34.0  # Uncomment for AWS Secrets Manager
# hvac==2.1.0    # Uncomment for HashiCorp Vault

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-swagger-ui-tag==0.6.8

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# File handling
python-magic==0.4.27
pillow==10.1.0

# Async utilities
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Configuration
dynaconf==3.2.4